Fix ability to click in canary and refactor
