+++ app/index.html
@@ -25,9 +25,6 @@
     <div id="main-app">
     </div>
 
-    <div id="app">
-    </div>
-
     <script src="assets/vendor.js"></script>
     <script src="assets/ember-twiddle.js"></script>
