+++ app/styles/_main.scss
@@ -8,7 +8,7 @@ body {
   width: 100%;
 }
 
-#app {
+#root {
   position: absolute;
 }
