+++ tests/index.html
@@ -21,9 +21,6 @@
     <div id="main-app">
     </div>
 
-    <div id="app">
-    </div>
-
     {{content-for "body"}}
     {{content-for "test-body"}}
